==> prep_scripts/2_VED.ipynb <==
import pandas as pd
import numpy as np

d1 = pd.read_csv("2_VED_orig_data.csv")
d1.head()

d1.columns, d1.shape

trips = np.unique(d1['trip_id'])
len(trips)

trips = np.unique(d1['trip_id'])
trips_tr = trips[:300]
trips_te = trips[300:]
tr = []
te = []
for j in range(d1.shape[0]):
    if d1['trip_id'][j] in trips_tr:
        tr.append(d1['trip_id'][j])
    else:
        te.append(d1['trip_id'][j])
#len(trips)
train_idx = np.array([x in trips[:300] for x in d1['trip_id'] ])
test_idx = np.array([x in trips[300:] for x in d1['trip_id']])
train_idx = np.argwhere(train_idx)[:,0]
test_idx = np.argwhere(test_idx)[:,0]
#sum(train_idx), sum(test_idx)

sum(train_idx), sum(test_idx)

len(tr),len(te)

X_labels = ['speed', 'speed_limit', 'speed_avg_week', 'speed_avg_time', 'speed_avg_week_time', 'speed_avg',
            'seconds', 'air_temperature',
       'wind_direction', 'wind_speed_ms', 'segangle',  
       'time', 'weekend',
       'drifting', 'dry', 'fog', 'freezing', 'none', 'snow', 'thunder', 'wet', 
        'living_street', 'motorway', 'motorway_link', 'primary', 'residential', 
       'secondary', 'secondary_link', 'service', 'tertiary', 
       'track', 'trunk', 'trunk_link', 'unclassified', 'unpaved']
y_labels = ['trip_id', 'trip_segmentno', 'segmentkey',
       'segmentid', 'ev_kwh']

d1.loc[train_idx, X_labels].to_csv("X_train_VED.csv", index=False)
d1.loc[train_idx, y_labels].to_csv("y_train_VED.csv", index=False)

d1.loc[test_idx, X_labels].to_csv("X_test_VED.csv", index=False)
d1.loc[test_idx, y_labels].to_csv("y_test_VED.csv", index=False)

tes = d1.loc[train_idx, y_labels]
tra = d1.loc[test_idx, y_labels]
len(np.unique(tra['trip_id'])), len(np.unique(tes['trip_id']))

tra.shape, tes.shape

12244, 7104
